+++ packages/expo-updates/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### 🎉 New features
 
 - Add rollback to embedded update directive. ([#21007](https://github.com/expo/expo/pull/21007), [#21652](https://github.com/expo/expo/pull/21652) by [@wschurman](https://github.com/wschurman))
+- New Updates.Provider JS API. ([#21880](https://github.com/expo/expo/pull/21880) by [@douglowder](https://github.com/douglowder))
 
 ### 🐛 Bug fixes